requirements.txt
aiogram>=3.2.0,<4.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
filelock>=3.12.0
